requirements.txt
pyspark
Faker
apache-airflow
kafka-python
kafka
faker
confluent_kafka
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
six==1.16.0
urllib3==2.1.0
cassandra-driver~=3.29.0